Rename assets related to extensions page from extensions to home
