+++ assets/home/home.scss
@@ -22,7 +22,7 @@ $break-medium: 783px; // adminbar goes big
 	src: url('../css/fonts/Reforma1918-Blanca.otf') format('opentype');
 }
 
-.sensei-extensions-page {
+.sensei-home-page {
 	margin-right: 10px;
 
 	@media (min-width: 783px) {
