+++ assets/home/update-notification/update-available.js
