+++ assets/home/update-notification/woocommerce-notice.js
