+++ webpack.config.js
@@ -47,8 +47,8 @@ const files = [
 	'js/stop-double-submission.js',
 	'setup-wizard/index.js',
 	'setup-wizard/style.scss',
-	'extensions/index.js',
-	'extensions/extensions.scss',
+	'home/index.js',
+	'home/home.scss',
 	'shared/styles/wp-components.scss',
 	'shared/components/modal/style.scss',
 	'data-port/import.js',
@@ -87,7 +87,7 @@ const files = [
 	'css/enrolment-debug.scss',
 	'css/frontend.scss',
 	'css/admin-custom.scss',
-	'css/extensions.scss',
+	'css/home.scss',
 	'css/global.scss',
 	'css/jquery-ui.css',
 	'css/modules-admin.css',
